requirements.txt
requests==2.30.0
pydantic==1.10.7
python-dotenv==0.19.2
pandas==2.0.1